moved paths and copyFile fn to utils, tweaked method calls
