+++ lib/new.js
@@ -3,17 +3,13 @@
 var fs = require('fs');
 var path = require('path');
 var config = require('./config');
-
-var paths = {
-    src: './lib/src/',
-    config: '_config.json',
-};
+var utils = require('./utils');
 
 function copySrcDirectories() {
-    var files = fs.readdirSync(paths.src);
+    var files = fs.readdirSync(utils.paths.src);
     files.forEach(function (file) {
-        paths.src = '/lib/src/';
-        var filePath = path.resolve(process.cwd() + paths.src + file);
+        utils.paths.src = '/lib/src/';
+        var filePath = path.resolve(process.cwd() + utils.paths.src + file);
         console.log(filePath);
 
         if (fs.statSync(filePath).isDirectory()) {
@@ -30,28 +26,16 @@ function copySrcDirectories() {
 function copyContentsFromDirectory(src, dest) {
     var files = fs.readdirSync(src);
     files.forEach(function (file) {
-        copyFile(path.resolve(src + '/' + file), path.resolve(dest + '/' + file));
-    });
-}
-
-function copyFile(src, dest) {
-    var data = fs.readFileSync(src, 'utf-8');
-
-    fs.writeFile(dest, data, function (err) {
-        if (err) return err;
+        utils.copyFile(path.resolve(src + '/' + file), path.resolve(dest + '/' + file));
     });
 }
 
 function newSite() {
-    var configSrc = paths.src + paths.config;
-    var configDest = path.resolve(process.cwd() + '/' + paths.config);
+    var configSrc = utils.paths.src + utils.paths.config;
+    var configDest = path.resolve(process.cwd() + '/' + utils.paths.config);
 
-    copyFile(configSrc, configDest);
+    utils.copyFile(configSrc, configDest);
     copySrcDirectories();
 }
 
-module.exports = {
-    newSite: newSite,
-    copyFile, copyFile,
-    paths: paths
-};
+module.exports = newSite;
